+++ .env
@@ -57,5 +57,12 @@ VITE_PUSHER_PORT="${PUSHER_PORT}"
 VITE_PUSHER_SCHEME="${PUSHER_SCHEME}"
 VITE_PUSHER_APP_CLUSTER="${PUSHER_APP_CLUSTER}"
 
+MIDTRANS_SERVER_KEY =SB-Mid-server-db8lMyVy_HmbfnQ_Dwl7udRa
+MIDTRANS_CLIENT_KEY =SB-Mid-client-j7PoL5uC89n-jiXV
+MIDTRANS_IS_PRODUCTION = false
+MIDTRANS_IS_SANITIZED = true
+MIDTRANS_IS_3DS =true
+
+
 WHATSAPP_TOKEN=EACBWqmZAc82ABOwxk43vZCJ1PSAQau1wuMAZBz2ZBNZAw8PIFfwfjf5xHt9frZCABUXC6Kbo7ZBDJcPltZC7MR38zkKrTJna9yc47jiVUUFFakI2etQaburWbxwNfO3kPxfsFJM8sTcCwRA9xbZA3iYSLz9Ddjnje5mQdbCbVpHBi6CpGzROF7nmw0VF1wZANiWU17ORZB6PvZC6n0i2iUSJLLfXnSmKyoEsCKLXjw4ZD
 PORT=3306
